chore: moving the codebase into a `src/` directory structure
