+++ __tests__/cmds/login.test.js
@@ -1,7 +1,7 @@
 const nock = require('nock');
 const config = require('config');
-const configStore = require('../../lib/configstore');
-const cmd = require('../../cmds/login');
+const configStore = require('../../src/lib/configstore');
+const cmd = require('../../src/cmds/login');
 
 describe('rdme login', () => {
   beforeAll(() => nock.disableNetConnect());
