+++ src/cmds/docs/index.js
