+++ src/cmds/logout.js
