+++ src/cmds/oas.js
@@ -10,7 +10,7 @@ exports.position = 1;
 exports.args = [];
 
 exports.run = function () {
-  const cp = spawn(path.join(__dirname, '..', 'node_modules', '.bin', 'oas'), process.argv.slice(3), {
+  const cp = spawn(path.join(__dirname, '..', '..', 'node_modules', '.bin', 'oas'), process.argv.slice(3), {
     stdio: 'inherit',
   });
