+++ src/cmds/open.js
