+++ src/cmds/versions/delete.js
