+++ src/cmds/whoami.js
