+++ src/lib/configstore.js
@@ -1,4 +1,4 @@
 const Configstore = require('configstore');
-const pkg = require('../package.json');
+const pkg = require('../../package.json');
 
 module.exports = new Configstore(`${pkg.name}-${process.env.NODE_ENV || 'production'}`);
